Broke up the views file into its sub-parts.
