+++ kakao/src/main/kotlin/com/agoda/kakao/views/KButton.kt
@@ -1,0 +1,19 @@
+package com.agoda.kakao.views
+
+import android.support.test.espresso.DataInteraction
+import android.view.View
+import com.agoda.kakao.assertions.TextViewAssertions
+import com.agoda.kakao.builders.ViewBuilder
+import org.hamcrest.Matcher
+
+/**
+ * View with BaseActions and TextViewAssertions
+ *
+ * @see BaseActions
+ * @see TextViewAssertions
+ */
+class KButton : KBaseView<KButton>, TextViewAssertions {
+    constructor(function: ViewBuilder.() -> Unit) : super(function)
+    constructor(parent: Matcher<View>, function: ViewBuilder.() -> Unit) : super(parent, function)
+    constructor(parent: DataInteraction, function: ViewBuilder.() -> Unit) : super(parent, function)
+}
