+++ kakao/src/main/kotlin/com/agoda/kakao/views/KEditText.kt
@@ -1,0 +1,20 @@
+package com.agoda.kakao.views
+
+import android.support.test.espresso.DataInteraction
+import android.view.View
+import com.agoda.kakao.actions.EditableActions
+import com.agoda.kakao.assertions.EditableAssertions
+import com.agoda.kakao.builders.ViewBuilder
+import org.hamcrest.Matcher
+
+/**
+ * View with EditableActions and EditableAssertions
+ *
+ * @see EditableActions
+ * @see EditableAssertions
+ */
+class KEditText : KBaseView<KEditText>, EditableActions, EditableAssertions {
+    constructor(function: ViewBuilder.() -> Unit) : super(function)
+    constructor(parent: Matcher<View>, function: ViewBuilder.() -> Unit) : super(parent, function)
+    constructor(parent: DataInteraction, function: ViewBuilder.() -> Unit) : super(parent, function)
+}
