+++ kakao/src/main/kotlin/com/agoda/kakao/views/KEmptyAdapterItem.kt
@@ -1,0 +1,13 @@
+package com.agoda.kakao.views
+
+import android.support.test.espresso.DataInteraction
+
+/**
+ * Empty implementation of KAdapterItem
+ *
+ * Use this if you want to perform/assert on the root view of adapter item
+ *
+ * @param parent Matcher of the root view of adapter item
+ * @see KAdapterItem
+ */
+class KEmptyAdapterItem(parent: DataInteraction) : KAdapterItem<KEmptyAdapterItem>(parent)
