+++ kakao/src/main/kotlin/com/agoda/kakao/views/KImageView.kt
@@ -1,0 +1,19 @@
+package com.agoda.kakao.views
+
+import android.support.test.espresso.DataInteraction
+import android.view.View
+import com.agoda.kakao.assertions.ImageViewAssertions
+import com.agoda.kakao.builders.ViewBuilder
+import org.hamcrest.Matcher
+
+/**
+ * View with BaseActions and ImageViewAssertions
+ *
+ * @see BaseActions
+ * @see ImageViewAssertions
+ */
+class KImageView : KBaseView<KImageView>, ImageViewAssertions {
+    constructor(function: ViewBuilder.() -> Unit) : super(function)
+    constructor(parent: Matcher<View>, function: ViewBuilder.() -> Unit) : super(parent, function)
+    constructor(parent: DataInteraction, function: ViewBuilder.() -> Unit) : super(parent, function)
+}
