+++ kakao/src/main/kotlin/com/agoda/kakao/views/KRecyclerItemType.kt
@@ -1,0 +1,11 @@
+package com.agoda.kakao.views
+
+import android.view.View
+import org.hamcrest.Matcher
+
+/**
+ * For internal use. Don't use manually.
+ *
+ * Holds type and corresponding provider function
+ */
+class KRecyclerItemType<out T : KRecyclerItem<*>>(val provideItem: (Matcher<View>) -> T)
