+++ kakao/src/main/kotlin/com/agoda/kakao/views/KSnackbar.kt
@@ -1,0 +1,22 @@
+package com.agoda.kakao.views
+
+import android.support.design.widget.Snackbar
+import android.support.v7.widget.AppCompatButton
+import android.support.v7.widget.AppCompatTextView
+
+/**
+ * View with internal TextView and a Button
+ *
+ * @see Snackbar
+ */
+class KSnackbar : KBaseView<KSnackbar>({ isInstanceOf(Snackbar.SnackbarLayout::class.java) }) {
+    val text = KTextView {
+        isDescendantOfA { isInstanceOf(Snackbar.SnackbarLayout::class.java) }
+        isInstanceOf(AppCompatTextView::class.java)
+    }
+
+    val action = KButton {
+        isDescendantOfA { isInstanceOf(Snackbar.SnackbarLayout::class.java) }
+        isInstanceOf(AppCompatButton::class.java)
+    }
+}
