+++ kakao/src/main/kotlin/com/agoda/kakao/views/KSwipeRefreshLayout.kt
@@ -1,0 +1,20 @@
+package com.agoda.kakao.views
+
+import android.support.test.espresso.DataInteraction
+import android.view.View
+import com.agoda.kakao.actions.SwipeRefreshLayoutActions
+import com.agoda.kakao.assertions.SwipeRefreshLayoutAssertions
+import com.agoda.kakao.builders.ViewBuilder
+import org.hamcrest.Matcher
+
+/**
+ * View with SwipeRefreshLayoutActions and SwipeRefreshLayoutAssertions
+ *
+ * @see SwipeRefreshLayoutActions
+ * @see SwipeRefreshLayoutAssertions
+ */
+class KSwipeRefreshLayout : KBaseView<KSwipeRefreshLayout>, SwipeRefreshLayoutActions, SwipeRefreshLayoutAssertions {
+    constructor(function: ViewBuilder.() -> Unit) : super(function)
+    constructor(parent: Matcher<View>, function: ViewBuilder.() -> Unit) : super(parent, function)
+    constructor(parent: DataInteraction, function: ViewBuilder.() -> Unit) : super(parent, function)
+}
