+++ kakao/src/main/kotlin/com/agoda/kakao/views/KTextInputLayout.kt
@@ -1,0 +1,38 @@
+package com.agoda.kakao.views
+
+import android.support.design.widget.TextInputEditText
+import android.support.test.espresso.DataInteraction
+import android.view.View
+import com.agoda.kakao.assertions.TextInputLayoutAssertions
+import com.agoda.kakao.builders.ViewBuilder
+import org.hamcrest.Matcher
+
+/**
+ * View with TextInputLayoutAssertions
+ *
+ * @see TextInputLayoutAssertions
+ */
+class KTextInputLayout : KBaseView<KTextInputLayout>, TextInputLayoutAssertions {
+    val edit: KEditText
+
+    constructor(function: ViewBuilder.() -> Unit) : super(function) {
+        edit = KEditText {
+            isDescendantOfA(function)
+            isInstanceOf(TextInputEditText::class.java)
+        }
+    }
+
+    constructor(parent: Matcher<View>, function: ViewBuilder.() -> Unit) : super(parent, function) {
+        edit = KEditText {
+            isDescendantOfA(function)
+            isInstanceOf(TextInputEditText::class.java)
+        }
+    }
+
+    constructor(parent: DataInteraction, function: ViewBuilder.() -> Unit) : super(parent, function) {
+        edit = KEditText {
+            isDescendantOfA(function)
+            isInstanceOf(TextInputEditText::class.java)
+        }
+    }
+}
