+++ kakao/src/main/kotlin/com/agoda/kakao/views/KView.kt
@@ -1,0 +1,18 @@
+package com.agoda.kakao.views
+
+import android.support.test.espresso.DataInteraction
+import android.view.View
+import com.agoda.kakao.builders.ViewBuilder
+import org.hamcrest.Matcher
+
+/**
+ * Simple view with BaseActions and BaseAssertions
+ *
+ * @see BaseActions
+ * @see BaseAssertions
+ */
+class KView : KBaseView<KView> {
+    constructor(function: ViewBuilder.() -> Unit) : super(function)
+    constructor(parent: Matcher<View>, function: ViewBuilder.() -> Unit) : super(parent, function)
+    constructor(parent: DataInteraction, function: ViewBuilder.() -> Unit) : super(parent, function)
+}
