+++ kakao/src/main/kotlin/com/agoda/kakao/views/KViewPager.kt
@@ -1,0 +1,20 @@
+package com.agoda.kakao.views
+
+import android.support.test.espresso.DataInteraction
+import android.view.View
+import com.agoda.kakao.actions.SwipeableActions
+import com.agoda.kakao.assertions.ViewPagerAssertions
+import com.agoda.kakao.builders.ViewBuilder
+import org.hamcrest.Matcher
+
+/**
+ * View with SwipeableActions and ViewPagerAssertions
+ *
+ * @see SwipeableActions
+ * @see ViewPagerAssertions
+ */
+class KViewPager : KBaseView<KViewPager>, SwipeableActions, ViewPagerAssertions {
+    constructor(function: ViewBuilder.() -> Unit) : super(function)
+    constructor(parent: Matcher<View>, function: ViewBuilder.() -> Unit) : super(parent, function)
+    constructor(parent: DataInteraction, function: ViewBuilder.() -> Unit) : super(parent, function)
+}
